requirements.txt
pip>=9
setuptools>=26
wheel>=0.29
pytest
coverage
python-gitlab
twine
pandas
numpy
matplotlib
xlrd
geopy
pmdarima
scikit-learn
statsmodels
holidays
streamlit
joblib
xmltodict
termcolor
google-cloud-storage
python-dotenv
gcsfs
